pyproject.toml
[project]
name = "testingfolder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.5.6",
    "fastapi>=0.135.3",
    "google-generativeai>=0.8.6",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.27",
    "langchain-google-genai>=4.2.1",
    "langchain-huggingface>=1.2.1",
    "langchain-pinecone>=0.2.13",
    "langchain-text-splitters>=1.1.1",
    "openai>=2.30.0",
    "pinecone>=7.3.0",
    "pypdf>=6.10.2",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "sentence-transformers>=5.4.0",
    "streamlit>=1.56.0",
    "tavily-python>=0.7.23",
    "uvicorn>=0.44.0",
]
